local mongo

+++ src/main/java/com/example/mongoRedis/auth/service/AuthServiceImpl.java
@@ -2,7 +2,6 @@ package com.example.mongoRedis.auth.service;
 
 import com.example.mongoRedis.auth.dto.request.LoginRequest;
 import com.example.mongoRedis.auth.dto.request.RefreshTokenRequest;
-import com.example.mongoRedis.auth.dto.request.SignupRequest;
 import com.example.mongoRedis.auth.dto.response.LoginResponse;
 import com.example.mongoRedis.auth.dto.response.RefreshTokenResponse;
 import com.example.mongoRedis.common.response.ApiResponse;
@@ -10,17 +9,14 @@ import com.example.mongoRedis.common.util.PasswordEncoderUtil;
 import com.example.mongoRedis.exception.CustomServiceException;
 import com.example.mongoRedis.jwt.JwtService;
 import com.example.mongoRedis.jwt.RedisTokenService;
-import com.example.mongoRedis.user.dto.model.Credentials;
 import com.example.mongoRedis.user.dto.model.User;
 import com.example.mongoRedis.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-
 @Service
 @RequiredArgsConstructor
-public class AuthServiceImpl implements AuthService{
+public class AuthServiceImpl implements AuthService {
 
     private final UserRepository userRepository;
     private final PasswordEncoderUtil passwordEncoder;
@@ -32,7 +28,11 @@ public class AuthServiceImpl implements AuthService{
         User user = userRepository.findByEmail(request.getEmail())
                 .orElseThrow(() -> new CustomServiceException("Invalid email or password"));
 
-        if (!passwordEncoder.matches(request.getPassword(), user.getCredentials().getPassword())) {
+        // if (!passwordEncoder.matches(request.getPassword(),
+        // user.getCredentials().getPassword())) {
+        // throw new CustomServiceException("Invalid email or password");
+        // }
+        if (!request.getPassword().equals(user.getCredentials().getPassword())) {
             throw new CustomServiceException("Invalid email or password");
         }
 
@@ -81,25 +81,30 @@ public class AuthServiceImpl implements AuthService{
 
     @Override
     public ApiResponse<Void> logout(String token) {
+        if (token != null && token.startsWith("Bearer ")) {
+            token = token.substring(7);
+        }
         String userId = jwtService.extractUserId(token);
-        redisTokenService.deleteAccessToken(userId);
-        redisTokenService.deleteRefreshToken(userId);
+        if (userId != null) {
+            redisTokenService.deleteAccessToken(userId);
+            redisTokenService.deleteRefreshToken(userId);
+        }
         return new ApiResponse<>(true, null, "Successfully logged out");
     }
 
-//    private User mapToUserEntity(SignupRequest request) {
-//        Credentials credentials = Credentials.builder()
-//                .username(request.getEmail())
-//                .password(passwordEncoder.encryptPassword(request.getPassword()))
-//                .roles(List.of("ROLE_" + request.getUserType().name()))
-//                .build();
-//
-//        return User.builder()
-//                .name(request.getName())
-//                .email(request.getEmail())
-//                .age(request.getAge())
-//                .userType(request.getUserType())
-//                .credentials(credentials)
-//                .build();
-//    }
+    // private User mapToUserEntity(SignupRequest request) {
+    // Credentials credentials = Credentials.builder()
+    // .username(request.getEmail())
+    // .password(passwordEncoder.encryptPassword(request.getPassword()))
+    // .roles(List.of("ROLE_" + request.getUserType().name()))
+    // .build();
+    //
+    // return User.builder()
+    // .name(request.getName())
+    // .email(request.getEmail())
+    // .age(request.getAge())
+    // .userType(request.getUserType())
+    // .credentials(credentials)
+    // .build();
+    // }
 }
